Change skill locks to use 16-bit data type
Fixes some issues when users had too many free skill locks.

+++ TSOClient/tso.client/UI/Panels/UIPersonPage.cs
@@ -328,7 +328,7 @@ namespace FSO.Client.UI.Panels
                         screen.vm.SendCommand(new VMNetSkillLockCmd()
                         {
                             SkillID = VMSkillMap[bar.SkillID],
-                            LockLevel = (byte)bar.LockLevel
+                            LockLevel = (short)bar.LockLevel
                         });
                     }
